==> notebooks/exploration/1.0-initial-processing-calls.ipynb <==
# package imports

import numpy as np
import pandas as pd

## Calls for Service Data

df_raw_calls_2012 = pd.read_csv('../data/raw/Calls_for_Service_2012.csv', usecols=[0,1,2,3,6,7,8,9,10,11,12,13,14,15])
df_raw_calls_2012['TimeCreate'] = pd.to_datetime(df_raw_calls_2012['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2012['TimeDispatch'] = pd.to_datetime(df_raw_calls_2012['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2012['TimeArrive'] = pd.to_datetime(df_raw_calls_2012['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2012['TimeClosed'] = pd.to_datetime(df_raw_calls_2012['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2012.set_index('NOPD_Item', inplace=True)

df_raw_calls_2013 = pd.read_csv('../data/raw/Calls_for_Service_2013.csv', usecols=[0,1,2,3,6,7,8,9,10,11,12,13,14,15])
df_raw_calls_2013['TimeCreate'] = pd.to_datetime(df_raw_calls_2013['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2013['TimeDispatch'] = pd.to_datetime(df_raw_calls_2013['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2013['TimeArrive'] = pd.to_datetime(df_raw_calls_2013['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2013['TimeClosed'] = pd.to_datetime(df_raw_calls_2013['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2013.set_index('NOPD_Item', inplace=True)

df_raw_calls_2014 = pd.read_csv('../data/raw/Calls_for_Service_2014.csv', usecols=[0,1,2,3,9,10,11,12,13,14,17,18,19,20])
df_raw_calls_2014['TimeCreate'] = pd.to_datetime(df_raw_calls_2014['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2014['TimeDispatch'] = pd.to_datetime(df_raw_calls_2014['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2014['TimeArrive'] = pd.to_datetime(df_raw_calls_2014['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2014['TimeClosed'] = pd.to_datetime(df_raw_calls_2014['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2014.set_index('NOPD_Item', inplace=True)

df_raw_calls_2015 = pd.read_csv('../data/raw/Calls_for_Service_2015.csv', usecols=[0,1,2,3,9,10,11,12,13,14,17,18,19,20])
df_raw_calls_2015['TimeCreate'] = pd.to_datetime(df_raw_calls_2015['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2015['TimeDispatch'] = pd.to_datetime(df_raw_calls_2015['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2015['TimeArrive'] = pd.to_datetime(df_raw_calls_2015['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2015['TimeClosed'] = pd.to_datetime(df_raw_calls_2015['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2015.set_index('NOPD_Item', inplace=True)

df_raw_calls_2016 = pd.read_csv('../data/raw/Calls_for_Service_2016.csv', usecols=[0,1,2,3,9,10,11,12,13,14,17,18,19,20])
df_raw_calls_2016['TimeCreate'] = pd.to_datetime(df_raw_calls_2016['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2016['TimeDispatch'] = pd.to_datetime(df_raw_calls_2016['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2016['TimeArrive'] = pd.to_datetime(df_raw_calls_2016['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2016['TimeClosed'] = pd.to_datetime(df_raw_calls_2016['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2016.set_index('NOPD_Item', inplace=True)

df_raw_calls_2017 = pd.read_csv('../data/raw/Calls_for_Service_2017.csv', usecols=[0,1,2,3,9,10,11,12,13,14,17,18,19,20])
df_raw_calls_2017['TimeCreate'] = pd.to_datetime(df_raw_calls_2017['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2017['TimeDispatch'] = pd.to_datetime(df_raw_calls_2017['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2017['TimeArrive'] = pd.to_datetime(df_raw_calls_2017['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2017['TimeClosed'] = pd.to_datetime(df_raw_calls_2017['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2017.set_index('NOPD_Item', inplace=True)

df_raw_calls_2018 = pd.read_csv('../data/raw/Calls_for_Service_2018.csv', usecols=[0,1,2,3,9,10,11,12,13,14,17,18,19,20])
df_raw_calls_2018['TimeCreate'] = pd.to_datetime(df_raw_calls_2018['TimeCreate'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2018['TimeDispatch'] = pd.to_datetime(df_raw_calls_2018['TimeDispatch'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2018['TimeArrive'] = pd.to_datetime(df_raw_calls_2018['TimeArrive'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2018['TimeClosed'] = pd.to_datetime(df_raw_calls_2018['TimeClosed'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_calls_2018.set_index('NOPD_Item', inplace=True)

df_calls = pd.concat([df_raw_calls_2012, df_raw_calls_2013, df_raw_calls_2014, df_raw_calls_2015, df_raw_calls_2016, df_raw_calls_2017, df_raw_calls_2018])

target_types = [
    '100',
    '100C',
    '100F',
    '100I',
    '100X',
    '1028P',
    '18',
    '18A',
    '18DE',
    '18F',
    '20',
    '20C',
    '20F',
    '20I',
    '20X',
    '98',
    '99']

target_dispositions = ['NAT', 'RTF', 'GOA']

df_traffic_calls = df_calls[df_calls['Type_'].isin(target_types)]

df_traffic_calls = df_traffic_calls[df_traffic_calls['Disposition'].isin(target_dispositions)]

## EPR Data

df_raw_epr_2012 = pd.read_csv('../data/raw/Electronic_Police_Report_2012.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2013 = pd.read_csv('../data/raw/Electronic_Police_Report_2013.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2014 = pd.read_csv('../data/raw/Electronic_Police_Report_2014.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2015 = pd.read_csv('../data/raw/Electronic_Police_Report_2015.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2016 = pd.read_csv('../data/raw/Electronic_Police_Report_2016.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2017 = pd.read_csv('../data/raw/Electronic_Police_Report_2017.csv', usecols=[0, 4, 7, 8, 20])

df_raw_epr_2018 = pd.read_csv('../data/raw/Electronic_Police_Report_2018.csv', usecols=[0, 4, 7, 8, 20])

df_epr = pd.concat([df_raw_epr_2012, df_raw_epr_2013, df_raw_epr_2014, df_raw_epr_2015, df_raw_epr_2016, df_raw_epr_2017, df_raw_epr_2018], ignore_index=True)

df_traffic_epr = df_epr[df_epr['Signal_Type'].isin(target_types)]

df_traffic_epr = df_traffic_epr[df_traffic_epr.Charge_Code.notna()]

## 311 Call Data

df_raw_311_calls = pd.read_csv('../data/raw/311_Calls__2012-Present_.csv', usecols=[0,1,2,3,4,5,6,7,8,11,14,15])
df_raw_311_calls['ticket_created_date_time'] = pd.to_datetime(df_raw_311_calls['ticket_created_date_time'], format='%m/%d/%Y %I:%M:%S %p')
df_raw_311_calls['ticket_closed_date_time'] = pd.to_datetime(df_raw_311_calls['ticket_closed_date_time'], format='%m/%d/%Y %I:%M:%S %p')

## Save Pickle Files

df_traffic_calls.to_pickle('../data/interim/traffic_calls.pickle')
df_traffic_epr.to_pickle('../data/interim/traffic_epr.pickle')
df_raw_311_calls.to_pickle('../data/interim/311_calls.pickle')
