==> notebooks/exploration/1.0-initial-processing-dpw.ipynb <==
import numpy as np
import pandas as pd

df_raw_dpw_projects = pd.read_csv('../../data/raw/Dept_of_Public_Works_Roadwork_Projects.csv', usecols=[1, 2, 3, 5, 8, 9])
df_raw_dpw_projects = df_raw_dpw_projects.drop_duplicates()
df_raw_dpw_projects['StartDate'] = pd.to_datetime(df_raw_dpw_projects.StartQtr, errors='coerce')
df_raw_dpw_projects['StartDate_2'] = pd.to_datetime(df_raw_dpw_projects[df_raw_dpw_projects.StartDate.isna()].StartQtr.str.replace('-...', ''))
df_raw_dpw_projects['StartDate'] = np.where(df_raw_dpw_projects.StartDate.isna(), df_raw_dpw_projects.StartDate_2, df_raw_dpw_projects.StartDate)
df_raw_dpw_projects['EndDate'] = pd.to_datetime(df_raw_dpw_projects.EndQtr, errors='coerce')
df_raw_dpw_projects['EndDate_2'] = pd.to_datetime(df_raw_dpw_projects[df_raw_dpw_projects.EndDate.isna()].EndQtr.str.replace('-...', ''))
df_raw_dpw_projects['EndDate'] = np.where(df_raw_dpw_projects.EndDate.isna(), df_raw_dpw_projects.EndDate_2, df_raw_dpw_projects.EndDate)
df_raw_dpw_projects = df_raw_dpw_projects.drop(columns=['StartQtr', 'EndQtr', 'StartDate_2', 'EndDate_2'])

df_raw_dpw_projects.to_pickle('../../data/interim/dpw_projects.pickle')
