==> notebooks/exploration/1.0-initial-processing-weather.ipynb <==
# package imports

import numpy as np
import pandas as pd

## Import Weather Data

df_raw_weather_lakefront = pd.read_csv('../data/raw/1399048.csv', usecols=[5, 8, 9, 35, 36, 38])
df_raw_weather_lakefront['DATE'] = pd.to_datetime(df_raw_weather_lakefront['DATE'])
df_raw_weather_lakefront['DAILYPrecip'] = pd.to_numeric(df_raw_weather_lakefront.DAILYPrecip.str.replace('[a-zA-Z]', ''), errors='coerce').fillna(0)
df_raw_weather_lakefront['HOURLYVISIBILITY'] = pd.to_numeric(df_raw_weather_lakefront.HOURLYVISIBILITY.str.replace('[a-zA-Z]', ''), errors='coerce').fillna(0)
df_raw_weather_lakefront['DAILYSunrise'] = pd.to_datetime(df_raw_weather_lakefront.DAILYSunrise.astype('str').str.pad(4, fillchar='0'), format='%H%M').dt.time
df_raw_weather_lakefront['DAILYSunrise'] = pd.to_datetime(df_raw_weather_lakefront.DATE.dt.date.astype('str') + ' ' + df_raw_weather_lakefront.DAILYSunrise.astype('str'))
df_raw_weather_lakefront['DAILYSunset'] = pd.to_datetime(df_raw_weather_lakefront.DAILYSunset, format='%H%M').dt.time
df_raw_weather_lakefront['DAILYSunset'] = pd.to_datetime(df_raw_weather_lakefront.DATE.dt.date.astype('str') + ' ' + df_raw_weather_lakefront.DAILYSunset.astype('str'))

df_raw_weather_lakefront.to_pickle('../data/interim/weather.pickle')
