==> notebooks/exploration/2.0-possible-outcomes.ipynb <==
# package imports
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import geopandas as gpd

# Accidents

1. DataFrame - one row per accident
2. Include Date/Time and Location

# import traffic pickle file
df_traffic = pd.read_pickle('../data/interim/traffic_calls.pickle')

# columns of interest
target_columns = [
    'Type_',
    'TypeText',
    'TimeCreate',
    'Location'
]

# select relevant columns
df_traffic = df_traffic[target_columns]

# select only accidents
df_accidents = df_traffic[(df_traffic.Type_.str.slice(0,2) == '20') | (df_traffic.Type_.str.slice(0,3) == '100')]

# split location into lat/lon
df_accidents['location_split'] = df_accidents.Location.str.replace('[()]', '').str.split(',')
df_accidents['lat'] = df_accidents.location_split.str.get(0).astype('float')
df_accidents['lon'] = df_accidents.location_split.str.get(1).astype('float')

# remove irrelevant columns
df_accidents.drop(columns=['Type_', 'TypeText', 'Location', 'location_split'], inplace=True)

# drop rows with missing lat/lon values
df_accidents.drop(df_accidents[df_accidents.lat <= 1].index, inplace=True)
df_accidents.drop(df_accidents[df_accidents.lat.isna()].index, inplace=True)

# save accidents dataframe to pickle file
df_accidents.to_pickle('../data/interim/accidents.pickle')

# Locations

1. Street Blocks
2. Census Tracts
3. Neighborhoods

## Street Blocks

# import road centerline shapefile
gdf_streets = gpd.read_file('../data/raw/Road_Centerline/geo_export_c02761a8-1d85-477e-a5cf-01b9f22f4d88.shp')

# target columns
street_columns = [
    'joinid',
    'fromright',
    'centerline',
    'roadclass',
    'fullname',
    'toleft',
    'toright',
    'fedrtetype',
    'fullnameab',
    'objectid',
    'strtetype',
    'fromleft',
    'geometry'
]

# select target columns
gdf_streets = gdf_streets[street_columns]

# set index to objectid
gdf_streets.set_index('objectid', inplace=True)

# write streets to pickle file
gdf_streets.to_pickle('../data/interim/streets.pickle')

## Census Tracts

# import Louisiana Census Tract file
gdf_tracts = gpd.read_file('../data/raw/tl_2017_22_tract/tl_2017_22_tract.shp')

# select Orleans Parish
gdf_tracts = gdf_tracts[gdf_tracts['COUNTYFP'] == '071']

# select target columns
tract_columns = [
    'GEOID',
    'geometry'
]

# remove extraneous columns
gdf_tracts = gdf_tracts[tract_columns].set_index('GEOID')

# save to pickle file
gdf_tracts.to_pickle('../data/interim/tracts.pickle')

## Neighborhoods

# import neighborhood shapefile
gdf_nhoods = gpd.read_file('../data/raw/Neighborhood_Statistical_Areas/Neighborhood_Statistical_Areas.shp')

# identify target columns
nhood_columns = [
    'OBJECTID',
    'GNOCDC_LAB',
    'geometry'
]

# remove extraneous columns
gdf_nhoods = gdf_nhoods[nhood_columns].set_index('OBJECTID')

# save to pickle file
gdf_nhoods.to_pickle('../data/interim/nhoods.pickle')

# Date/Time

1. 15-min increments
2. 30-min increments
3. 1-hr increments
4. 1-day increments

# set start and end datetimes for ranges
range_start = pd.Timestamp('2012-01-01 00:00:00')
range_end   = pd.Timestamp('2018-07-01 00:00:00')

# create interval indexes
date_range_15_min = pd.interval_range(start=range_start, end=range_end, freq='15min', closed='left')
date_range_30_min = pd.interval_range(start=range_start, end=range_end, freq='30min', closed='left')
date_range_60_min = pd.interval_range(start=range_start, end=range_end, freq='60min', closed='left')
date_range_1_day = pd.interval_range(start=range_start, end=range_end, freq='D', closed='left')
