==> notebooks/exploration/2.0-possible-predictors.ipynb <==
# package imports
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import geopandas as gpd

# Time-Based

1. Current/Previous Weather
2. Day/Night
3. Rush hour

# set start and end datetimes for ranges
range_start = pd.Timestamp('2012-01-01 00:00:00')
range_end   = pd.Timestamp('2018-07-01 00:00:00')

# create interval indexes
date_range_60_min = pd.interval_range(start=range_start, end=range_end, freq='60min', closed='left')
date_range_1_day = pd.interval_range(start=range_start, end=range_end, freq='D', closed='left')

# import weather pickle files
df_weather = pd.read_pickle('../data/interim/weather.pickle')
df_weather.set_index('DATE', inplace=True)

# create inclement_weather variable
df_weather['inclement_weather'] = np.where(df_weather.HOURLYPRSENTWEATHERTYPE.isna(), 0, 1)

# resample to group by hour and day
df_weather_hour = df_weather[['DAILYPrecip', 'inclement_weather']].resample('60T').max()
df_weather_day =  df_weather[['DAILYPrecip', 'inclement_weather']].resample('D').max()

df_hour = pd.DataFrame(index=date_range_60_min)
df_day = pd.DataFrame(index=date_range_1_day)

# Location Based

1. Road work
2. Road type

# Event Based

1. Previous accidents
2. Vicinity to traffic offenses

pd.period_range('2012-01-01 00:00:00', '2018-06-30 23:30:00', freq='30min')

gpd.read_file('../data/raw/Road_Centerline/geo_export_c02761a8-1d85-477e-a5cf-01b9f22f4d88.shp')
