==> notebooks/exploration/2.0-traffic-data.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

df_calls = pd.read_pickle('../data/interim/traffic_calls.pickle')

df_epr = pd.read_pickle('../data/interim/traffic_epr.pickle')

target_charges = [
    'RECKLESS OPERATION OF A VEHICLE',
    'REASONABLE SAFETY - BACKING',
    'ONE WAY ROADWAYS & TRAFFIC ISLANDS',
    'FLIGHT FROM AN OFFICER',
    'DISREGARDING STOP AND YIELD SIGNS',
    'DRIVING WHILE INTOXICATED',
    'CARELESS OPERATION OF A VEHICLE',
    'TURNING AT INTERSECTIONS',
    'EXCEEDING SAFE SPEED',
    'EXCEEDING SPEED LIMIT',
    'DRIVING ROADWAY LANED FOR TRAFFIC',
    'DISREGARDING TRAFFIC CONTROL SIGN',
    'DISREGARDING SIGNS AND SIGNALS',
    'DRIVING ON ROADWAY LANED  / TRAFFIC',
    'STOPPING AT STOP SIGNS',
    'HIT & RUN DRIVING',
    'DRIVING ON DIVIDED HIGHWAYS',
    'FOLLOWING TOO CLOSELY',
    'ESCAPE',
    'INTIMIDATION',
    'DRIVING TOO SLOWLY',
    'OVERTAKING ON LEFT',
    'IMPROPER USE OF TURN SIGNALS',
    'DRIVING ON DIVIDED STREET - RULES',
    'MUST YIELD TO EMERGENCY VEHICLE',
    'OBEDIENCE TO TRAFFIC CONTROL',
    'ONE WAY STREETS AND ALLEYWAYS',
    'DRIVING IN PROHIBITED AREAS',
    'OBEDIENCE TO TRAFFIC CONTROLS',
    'SPEED LIMITS',
    'MOVEMENT ON ONE WAY STREETS',
    'YIELDING TO EMERGENCY VEHICLES',
    'TRUCK ROUTE VIOLATIONS',
    'VEHICLES ENTITLED TO RIGHT OF WAY',
    'PROPER POSITION TO TURN',
    'SPEED LIMITS - SCHOOL ZONES',
    'OPENING VEHICLE DOOR IN TRAFFIC',
    'PUBLIC DRUNK; DRUG INCAPACITATION']

df_epr_hazards = df_epr[df_epr['Charge_Description'].isin(target_charges)].groupby('Item_Number').count()[['Charge_Description']]

df_epr_hazards.columns = ['hazard_count']

df_epr_hazards.index = df_epr_hazards.index.str.replace('-', '')

df_calls_epr = pd.merge(df_calls, df_epr_hazards, left_index=True, right_index=True, how='left')

df_calls_epr['hazard_count'] = df_calls_epr.hazard_count.fillna(0).astype('int')

df_incidents = df_calls_epr[['Type_', 'TimeCreate', 'Location', 'hazard_count']]

df_incidents[['lat', 'lon']] = df_incidents.Location.str.replace('[()]', '').str.split(',', expand=True)

df_incidents['lat'] = pd.to_numeric(df_incidents.lat)

df_incidents['lon'] = pd.to_numeric(df_incidents.lon)

df_incidents.drop(columns='Location', inplace=True)

df_calls[['Type_', 'TypeText']].drop_duplicates().set_index('Type_').sort_index()
