==> notebooks/exploration/3.0-bokeh-map.ipynb <==
from bokeh.io import show, output_notebook
from bokeh.plotting import figure
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.pylab as pylab
import seaborn as sns

import shapely as shp
import geopandas as gpd
from scipy import ndimage

from sklearn.model_selection import train_test_split

data = pd.read_pickle('../../data/processed/all_samples.pickle')
data['datetime'] = pd.to_datetime(data.date)
data['day'] = data.datetime.dt.weekday_name

features = ['hour',
            'daylight_yn',
            'holiday_yn',
            'rush_hour_yn',
            'temp',
            'wind_speed',
            'precipitation',
            'road_length',
            'class_freeway',
            'class_local',
            'class_major',
            'class_other',
            'class_unimproved',
            'day_Monday',
            'day_Tuesday',
            'day_Wednesday',
            'day_Thursday',
            'day_Friday',
            'day_Saturday',
            'day_Sunday']

labels = 'accident_yn'

X = data[features]
y = data[labels]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

import pickle
f = open('../../data/processed/rf_random.pickle', 'wb')
pickle.dump(rf_random, f)
f.close()
best_random = rf_random.best_estimator_
predictions = model.predict(test_features)
random_performance = evaluate(best_random, X_test, y_test)

import pickle
f = open('../../data/processed/rf_random.pickle', 'rb')
rf = pickle.load(f)
f.close()

rf.best_estimator_.predict

output_notebook()

## Maps

streets = gpd.read_file('../../data/raw/Road_Centerline/geo_export_c02761a8-1d85-477e-a5cf-01b9f22f4d88.shp')
street_columns = ['joinid', 'roadclass', 'fullname', 'geometry']
streets = streets[street_columns]
streets = streets.dissolve(by='joinid', aggfunc='first')
streets['roadclass'] = streets.roadclass.fillna('Other')

neighborhoods = gpd.read_file('../../data/raw/Neighborhood_Statistical_Areas/Neighborhood_Statistical_Areas.shp')
nhood_columns = ['OBJECTID', 'GNOCDC_LAB', 'geometry']
neighborhoods = neighborhoods[nhood_columns]
neighborhoods.columns = ['nhood_id', 'nhood', 'geometry']
neighborhoods.set_index('nhood_id', inplace=True)
neighborhoods = neighborhoods.to_crs(epsg=4326)

sample_data = pd.read_pickle('../../data/processed/all_samples.pickle')
sample_data['datetime'] = pd.to_datetime(sample_data.date)

accidents = pd.read_pickle('../../data/interim/features/accidents.pickle')
accidents['geometry'] = list(zip(accidents.lon, accidents.lat))
accidents['geometry'] = accidents.geometry.apply(shp.geometry.Point)
accidents = gpd.GeoDataFrame(accidents, geometry='geometry')
accidents.crs = {'init' :'epsg:4326'}

figure_size = (20, 20)

nhood_base = neighborhoods.plot(figsize=figure_size,
                                color='gray',
                                edgecolor='black',
                                alpha=.5)

streets_local = streets[streets['roadclass'] == 'Local'].plot(ax=nhood_base,
                                                              figsize=figure_size,
                                                              color='white',
                                                              alpha=.2)

streets_major = streets[streets['roadclass'] == 'Major Arterial'].plot(ax=streets_local,
                                                                       figsize=figure_size,
                                                                       color='green',
                                                                       alpha=.4)

streets_freeway = streets[streets['roadclass'] == 'Freeway'].plot(ax=streets_major,
                                                                  figsize=figure_size,
                                                                  color='blue',
                                                                  alpha=.5)

nhood_base = neighborhoods.plot(figsize=figure_size,
                                color='gray',
                                edgecolor='gray',
                                alpha=.5)

accidents.plot(ax=nhood_base, marker='*', color='red', markersize=1, alpha=.2)

plot = figure(plot_width=400, tools='pan,box_zoom')
plot.circle([1,2,3,4,5], [8,6,5,2,3])
show(plot)
