==> notebooks/exploration/3.0-prediction-results-streets.ipynb <==
import numpy as np
import pandas as pd
import pickle
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn.ensemble import RandomForestClassifier
from pprint import pprint

data = pd.read_pickle('../../data/processed/all_samples.pickle')
data['datetime'] = pd.to_datetime(data.date)
data['day'] = data.datetime.dt.weekday_name
data = pd.get_dummies(data, prefix='day', columns=['day'])

features = ['hour',
            'daylight_yn',
            'holiday_yn',
            'rush_hour_yn',
            'temp',
            'wind_speed',
            'precipitation',
            'road_length',
            'class_freeway',
            'class_local',
            'class_major',
            'class_other',
            'class_unimproved',
            'day_Monday',
            'day_Tuesday',
            'day_Wednesday',
            'day_Thursday',
            'day_Friday',
            'day_Saturday',
            'day_Sunday']

labels = 'accident_yn'

X = data[features]
y = data[labels]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=42)

f = open('../../data/processed/rf_random.pickle', 'rb')
rf = pickle.load(f)
f.close()

model = rf.best_estimator_

df = X_test.copy()

df['predicted'] = model.predict(X_test)

results = df.join(y_test)

df_compare = results.join(data[['segment_id', 'datetime']])

target_columns = ['segment_id',
                  'datetime',
                  'hour',
                  'daylight_yn',
                  'temp',
                  'wind_speed',
                  'predicted',
                  'accident_yn']

df_output = df_compare[target_columns].reset_index(drop=True)

rename_columns = ['segment_id',
                  'date',
                  'hour',
                  'daylight_yn',
                  'temp',
                  'wind_speed',
                  'predicted',
                  'actual']

df_output.columns = rename_columns

street_info = pd.read_pickle('../../data/interim/features/streets_by_nhood.pickle')

street_info = street_info.set_index('segment_id')

street_class = street_info[['class_Freeway', 'class_Local', 'class_Major Arterial', 'class_Other', 'class_Umimproved']].idxmax(axis=1)

street_class.name = 'class'

streets = street_info.join(street_class)

streets = streets[['fullname', 'class', 'shape_leng', 'geometry']]

streets['class'] = streets['class'].str[6:]

streets

# True Positive
df_output.loc[(df_output['predicted'] == 1) & (df_output['actual'] == 1), 'prediction'] = 'True Positive'

# False Positive
df_output.loc[(df_output['predicted'] == 1) & (df_output['actual'] == 0), 'prediction'] = 'False Positive'

# False Negative
df_output.loc[(df_output['predicted'] == 0) & (df_output['actual'] == 1), 'prediction'] = 'False Negative'

# True Negative
df_output.loc[(df_output['predicted'] == 0) & (df_output['actual'] == 0), 'prediction'] = 'True Negative'

# True Positive
df_output['tp'] = np.where((df_output['predicted'] == 1) & (df_output['actual'] == 1), 1, 0)

# False Positive
df_output['fp'] = np.where((df_output['predicted'] == 1) & (df_output['actual'] == 0), 1, 0)

# False Negative
df_output['fn'] = np.where((df_output['predicted'] == 0) & (df_output['actual'] == 1), 1, 0)

# True Negative
df_output['tn'] = np.where((df_output['predicted'] == 0) & (df_output['actual'] == 0), 1, 0)

df_output.head()

street_accident_predictions = pd.merge(streets, df_output, how='left', on='segment_id')

street_accident_predictions.columns

street_date_groups = street_accident_predictions[['segment_id', 'date', 'tp', 'fp', 'fn', 'tn']]

street_date_groups['year'] = street_date_groups.date.dt.year
street_date_groups['month'] = street_date_groups.date.dt.month

street_date_groups = street_date_groups[['segment_id', 'tp', 'fp', 'fn', 'tn', 'year', 'month']]

street_predictions = street_date_groups[['segment_id', 'tp', 'fp', 'fn', 'tn']].groupby('segment_id').sum()

street_output = streets.join(street_predictions)

street_output = street_output.reset_index()

street_output.to_pickle('../../data/processed/street_predictions.pickle')
